python/requirements.txt
rembg==2.0.75
onnxruntime>=1.17
pillow>=10
